+++ src/test/scala/scorex/lagonaki/integration/api/UtilsAPISpecification.scala
@@ -9,11 +9,15 @@ class UtilsAPISpecification extends FunSuite with Matchers {
 
   import scorex.lagonaki.TestingCommons._
 
-  test("/seed API route") {
+  test("/utils/hash API route") {
+  //TODO
+  }
+
+  test("/utils/seed API route") {
     Base58.decode((getRequest("/utils/seed") \ "seed").as[String]).isSuccess shouldBe true
   }
 
-  test("/seed/{length} API route") {
+  test("/utils/seed/{length} API route") {
     val length = Random.nextInt(4096)
     Base58.decode((getRequest(s"/utils/seed/$length") \ "seed").as[String]).get.length shouldBe length
   }