+++ scorex-transaction/src/main/scala/scorex/api/http/AddressApiRoute.scala
@@ -109,7 +109,7 @@ case class AddressApiRoute(override val application: Application)(implicit val c
       value = "Json with data",
       required = true,
       paramType = "body",
-      dataType = "SignedMessage",
+      dataType = "scorex.api.http.SignedMessage",
       defaultValue = "{\n\t\"message\":\"Plain message\",\n\t\"signature\":\"Base58-encoded signature\",\n\t\"publickey\":\"Base58-encoded public key\"\n}"
     )
   ))
@@ -283,11 +283,11 @@ case class AddressApiRoute(override val application: Application)(implicit val c
 
 
   private def verifyPath(address: String, decode: Boolean) = {
-    withCors {
-      entity(as[String]) { jsText =>
-        complete {
+    entity(as[String]) { jsText =>
+      withAuth {
+        postJsonRoute {
           val parsed = Try(Json.parse(jsText)).getOrElse(WrongJson.json)
-          val jsRes = parsed.validate[SignedMessage] match {
+          parsed.validate[SignedMessage] match {
             case err: JsError =>
               WrongJson.json
             case JsSuccess(m: SignedMessage, _) =>
@@ -308,7 +308,6 @@ case class AddressApiRoute(override val application: Application)(implicit val c
                 }
               }
           }
-          Json.stringify(jsRes)
         }
       }
     }